cache packing for crafting
