+++ content/skills/crafting/pack/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+    id("base-conventions")
+}
