+++ content/skills/crafting/pack/src/main/resources/pack/configs/crafting_npcs.toml
@@ -1,0 +1,23 @@
+[[npc]]
+id = "npc.ellis_tanner"
+inherit = "npc.ellis_tanner"
+moveRestrict = "Indoors"
+wanderRange = 5
+
+[[npc]]
+id = "npc.master_crafter"
+inherit = "npc.master_crafter"
+moveRestrict = "NoMove"
+wanderRange = 0
+
+[[npc]]
+id = "npc.master_crafter_2"
+inherit = "npc.master_crafter_2"
+moveRestrict = "NoMove"
+wanderRange = 0
+
+[[npc]]
+id = "npc.master_crafter_3"
+inherit = "npc.master_crafter_3"
+moveRestrict = "NoMove"
+wanderRange = 0
