+++ or-cache/src/main/kotlin/dev/openrune/CacheTools.kt
@@ -36,7 +36,6 @@ import dev.openrune.tables.consumables.food.FoodTable
 import dev.openrune.tables.consumables.potion.PotionEffectTable
 import dev.openrune.tables.consumables.potion.PotionTable
 import dev.openrune.tables.skills.Cooking
-import dev.openrune.tables.skills.Crafting
 import dev.openrune.tables.skills.Firemaking
 import dev.openrune.tables.skills.Herblore
 import dev.openrune.tables.skills.Mining
@@ -186,11 +185,6 @@ fun tablesToPack(): List<DBTable> = listOf(
     Herblore.barbarianMixes(),
     Herblore.swampTar(),
     Herblore.crushing(),
-    Crafting.facilities(),
-    Crafting.hand(),
-    Crafting.tanning(),
-    Crafting.silver(),
-    Crafting.gold(),
     Smithing.bars(),
     Smithing.cannonBalls(),
     Smithing.dragonForge(),
